ReceiveCommits: Add thread-safety requirement as JavaDoc to sendMessages method
AsyncReceiveCommits may call ReceiveCommits#sendMessages while
ReceiveCommits#processCommands is still running (if the execution of
processCommands takes too long and AsyncReceiveCommits gets a timeout).
This means that local variables that are accessed in this method must be
thread-safe (otherwise we may hit a ConcurrentModificationException if
we read a variable in this method that at the same time is updated by
the background thread that still executes processCommands).

Fixes for ConcurrentModificationExceptions have been made by change
I5816fec63 and change I8b6e5de2a. Documenting this as JavaDoc should
help us to avoid similar issues in future.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I93cf5dc347e85bfd1264d3268e397cc53d53679c

Former-commit-id: cf912c3496184264008891c13850fdacdb94db12