Cache ranges for Highlight.js syntax markers
The cache is saving 5-10% time for syntax layer. Since the syntax layer
process whole files (not only diff chunks) there is a lot of duplicity
between original and modified file (left and right diff). Calculation
of ranges is an quite expensive operation, so any saving for skipping
calculation is improving the time of javascript execution.

Change-Id: I0976f25238c452a1240e96e4631e0aa9686dfa67

Former-commit-id: eada9d2819ce11c09b95689ee3c3a3092a8ffba1