Add additional Cherry-Pick method
There already existed 2 cherry-pick methods: One for cherry-picking a
commit and another for cherry-picking a change.

This change adds a third method that simply add more parameters such as

topic: Name of the topic to set on the cherry-picked change.

revertedChange: Id that could be set as the RevertOf of the
cherry-picked change, and also ensures the message is consistent with
reverting a change (therefore no cherry-pick message).

changeIdForNewChange: Sets the cherry-picked change to a specific
changeId.

It was possible to move this function to a utility class such as
CommitUtil, but it was not done since CherryPickChange is already sort
of a utility class (as it is not a REST endpoint class).

Change-Id: Iee3f3648330883c9e5ed5c5c075879f316f5783e

Former-commit-id: be97c9d344dbec296366798b122d0dc92bb50ed5