Update rules_closure to latest version
The latest version removed Java 7 compatibility mode [1] for
protobuf_java rule.

Note that this above fix is only related to the building protobuf_java
rule itself, and is not related to the building of java_proto_library,
that is also affected, but for another reason: default java toolchain
in currently used java_tools distribution is also targeting Java
language level 7 for proto rules. Luckily my PR was accepted here and
Java 7 compatibility mode was removed. However, to take effect a new
java_tools version must be released and Bazel must be updated to point
to it. This will be done in the follow-up change.

[1] https://github.com/bazelbuild/rules_closure/pull/432

Bug: Issue 11102
Change-Id: I0860772dbf78a397f1f8f78f17eb8dcec4ac07c0

Former-commit-id: 39f4f5b20c77e5344fef895baa155d57bd27a32b