Emails: Add toUserIdentity method to reduce duplicate code
The EventFactory and PatchSetInfoFactory classes both have exactly the
same method, and there is a TODO to find a common place for it.

Given that the method implementation uses Emails.getAccountFor, and
both EventFactory and PatchSetInfoFactory are already injecting Emails,
it makes sense to just move the method to Emails rather than creating
yet another utility class or adding it to some other existing utility
class where it doesn't really fit.

Change-Id: If8c90b0cdf489e3a627e312111dd17b1d1b85240

Former-commit-id: 1a2cb983cd2f847bf53e6bd42ac493b8031122b3