restapi: add an alternative CreateChange endpoint
In distributed installations, such as googlesource.com, different
tasks (running on different machines) are responsible for different
projects. This new API endpoint has the project name in the URL, so
load balancers can route requests to the task that has the data in
memory.

It might be more principled to add the whole /changes/ collection as a
child to /projects/ , but this is hard because the changes collection
is hard coded to be a a toplevel resource.

Every other REST operations on changes includes a change identifier
(eg. project~123), and can be routed based on that identifier.

Change-Id: I318603de0418e177f742684567867bf90571ac70

Former-commit-id: 44c724e185247d075500212b29ba0d009b2d2011