Replace references to bazel-genfiles with bazel-bin
In bazel 0.25 the target directory was changed from bazel-genfiles
to bazel-bin and bazel-genfiles was symlinked to bazel-bin.

In bazel 1.0 the symlink no longer gets created, and this causes the
./tools/maven/api.sh script to fail on a clean workspace because it
still refers to the old location:

  ./tools/maven/api.sh: line 66:
  ./bazel-genfiles/tools/maven/api_install.sh: No such file or directory

Fix api.sh to use bazel-bin, and also update dev-build-plugin.txt to
refer to bazel-bin.

Note that some documentation was previously updated in I9521cd7e5.

Bug: Issue 11757
Change-Id: I21f204d44ac78559f51ad2e62b085240170d83a7

Former-commit-id: 1c04c59e80838a9c3d72bfbabdb5d64cd0ac5c55