Add idForNewChange parameter for cherry-picking a change
Add capability when performing the cherry-pick change method to
specify the numeric change-id of the new change that is to be created.
Cherry-pick is also possible to an existing change, and it will just
create a new patchset; in this case, we check to see that the inputs
given to the function agree with the Change-Id and the Id of that
change, otherwise we throw an error.

This change is useful when performing a batchUpdate. When performing a
batchUpdate it is necessary to know the numeric change id in advance, so
without passing the numeric change id as a parameter it will not be
possible to perform a batchUpdate with both the cherry-pick and any
other update that involves the newly created change (e.g sending
notifications).

Change-Id: I156d30ff719106041f499f0e73b456478f1e1e97

Former-commit-id: 777ff4715d92eaaef5e149334d901188359f3294