Remove unused classes to read/write legacy change notes
The comments in the change notes are nowadays always stored as JSON. The
classes to read/write legacy change notes were only kept because Google
didn't do the migration to the JSON format yet. However this migration
is long done by now and when the cleanup after this migration was done
(see change Iadc371e0f) it was forgotten to remove these classes as
well.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I8c6ac4972c76cf1687e3902f50378f6d6031b0c1

Former-commit-id: 1dfc3e58e5eab52e595ec1353589182db7309c29