When publishing draft comments ignore comments on non-existing patch sets
It can happen that we see a draft comment for on existing patch set if
changes with the same numeric ID exist:

- change 12345 has 3 patch sets in repo X
- another change 12345 has 7 patch sets in repo Y
- the user saves a draft comment on patch set 6 of the change in repo Y
- this draft comment gets stored in:
  AllUsers -> refs/draft-comments/45/12345/<account-id>
- when posting a review with draft handling PUBLISH_ALL_REVISIONS on the
  change in repo X, the draft comments are loaded from
  AllUsers -> refs/draft-comments/45/12345/<account-id>, including the
  draft comment that was saved for patch set 6 of the change in repo Y
- patch set 6 does not exist for the change in repo x, hence that patch
  set cannot be found

Instead of failing hard (and returning an Internal Server Error) to the
caller, just ignore that comment.

Gerrit ensures that numeric change IDs are unique, but you can get
duplicates if change refs of one repo are copied/pushed to another repo
on the same host (this should never be done, but we know it happens).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I7a27b09ba3ab23fb9f111818d9ac8b9cc592cfdc

Former-commit-id: e028fa3a3e345bd97576f82f4f4d34c442f02710