MergeOp: Log IntegrationExceptions that are treated as conflict only as warning
We convert IntegrationExceptions into ResourceConflictExceptions. This
means for the caller this is a 409 Conflict and not an error. If it's
not an error we should not log it at severe level. For example such an
IntegrateException occurs if a change cannot be submitted because there
is a conflict with the destination branch.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I0ae5186fae346c71bad473092bb554e43c03515e

Former-commit-id: 57128499260be933741f61cd017c48ca44464e10