Use ReceivePackRefCache in all places where ReceiveCommits needs refs
This commit removes the usage of ReceivePack#getAdvertisedRefs in favor
of the newly created ReceivePackRefCache.

This is mostly a cleanup, but for ReplaceRequest#validateNewPatchSetNoteDb
it will provide a real speed-up as the operation moves from O(refs) to
O(1).

Before and after this commit, all call sites use(d) the permission-aware
ref database.

Change-Id: I325099a045c1f77db0ab525bc2bc814f1524a8b3

Former-commit-id: 95c72b27c7dc9e92ccd02aa6d942b944af255d74