Lazyload cursor stops once dropdown is opened
We don't need cursor stops to be calculated when dropdown is not
opened. Their calculation is expensive, so with this lazy loading,
we should get a nice perf improvement (~50ms) on app start. There is
also fix for gr-dropdown to works when user press enter, which got
broken by moving to polymer 2 and start using shadow dom.

Change-Id: Ib81b4d4dc6001605e240a33554adc13a24193b8c

Former-commit-id: 7f9abf6165c229b82e76fa078be0cc3038949cb8