Add REST endpoint to list label definitions of a project
Review labels are defined in the project.config file inside the
refs/meta/config branch. Viewing and updating the label definitions of
a project is currently only possible via git access to the
refs/meta/config branch.

We want to allow editing review labels from the UI. For this the backend
must provide REST endpoints to list and update the labels of a project.

This change implements a new REST endpoint that allows listing the label
definitions of one project. Inherited label definitions are not returned
yet, but a follow-up change should add an option on the List Labels REST
endpoint to include them. This is why in the Java extension API the
labels() method returns a ListLabelsRequest instead of directly the list
of labels. The ListLabelsRequest will allow callers to set options on
the request before executing it.

Planned follow-up changes:
- Add option on List Labels REST endpoint to include inherited labels
- Add REST endpoint to get a single label
- Add REST endpoint to update an existing label
- Add REST endpoint to add a new label
- Add REST endpoint to remove a label
- Implement a UI for editing labels

Bug: Issue 11522
Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ia65760bbf893ea81747627fcdd0306f6e750f4dd

Former-commit-id: 44098d0219a5c4ca232c6c865b0c59698d469557