Improve error message for anonymous calls to label REST endpoints
At the moment anonymous requests fail with an AuthException saying that
read/write permissions for refs/meta/config are missing. It's better to
make it explicit that the request is rejected because it was anonymous,
as it happens often that users are unaware that they made an anonymous
call and then they wonder why it complains about permissions that they
actually have.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I10a393813210ac8c57307d1781814a8f3dd88e85

Former-commit-id: af3e24d6bcc6bcdf4195695da7efe4734bf375e8