OutgoingEmail: Sanity check recipients of multipart and plaintext
The email gets sent twice; once to users who have requested HTML/multipart
and again to users who have requested plain text.

Add a sanity check and log an error if there is any intersection between
the two sets of users. If this is the case it means those users will get
the mail twice.

Change-Id: Id8fafe335872235a84c0f34905a7ad3215ad0c2f

Former-commit-id: fd8b0f2ff2bb3377985d75be850f96022f2b8b02