Keep only AccountLoader.Factory in REST handler
This is a follow up to I41071ded4. It's cleaner to keep only factories
in REST API handlers - even when the class is not a Singleton. This way,
copy-and-paste errors are reduced to a minimum.

An ideal solution would check that @Singletons reference only other
Singletons or factories and providers.

Change-Id: Ieae81dbc92b9546bbe12af29f3040362e5eb5f96

Former-commit-id: 655529390b4e3756c2759461f9c8a3cf7a268d81