Elasticsearch: Don't use JsonElement.toString
JsonElement.toString returns the value wrapped in an extra pair of
quotes, while getAsString does not.

The base64 decoder that we are currently using seems to be tolerant
of this, but others may not be which could cause issues.

Also, we already use getAsString in other places, so changing this
makes it consistent.

Change-Id: I42c71444782aa467321027856b017df62106c38c

Former-commit-id: 419bdcc59a94b3fa4b4c5c439389ef03d9d190e4