Expose query submission ID to the user
As a follow-up for RevertSubmission (91861e8f03 [formerly b81f56c469]), the users
could wish to know the changes of a certain submission if they
are considering to revert the entire submission, or if the submission
was already reverted and they wish to find it quickly.

Also, it is important for the frontend to know which changes belong
to a certain submission, so that they can suggest a better commit
message to the user when the user wants to revert a submission.

Change-Id: I337f7f8aeea2bc4146ed7e642f41c8710a153634

Former-commit-id: 63f23e3b82eec150f3b310fd7a7f615feebf8a5d