Fix formatting of submission IDs
When changes get submitted we use the RequestId as submission ID, but
the method that formatted the RequestId for storage was cutting off the
first and last characters from the RequestId. Earlier this was needed
because the RequestId string included square brackets around the actual
request ID. Change I116a188cf removed the square brackets but forgot to
adapt the toStringForStorage() method.

Change-Id: I2d46ede7096eec796a6d99987b7dfff20c3bcb33
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: e3c248c8852bbe4aa047b4db576780b46fcea257