Fix reporting for flaky builds in Jenkinsfile
The verification pipeline was not properly reporting the results of
builds that were suspect to being flaky. This happened, because the
build did not save the results in a variable, if the propagate option
was used, which had to be done for the pipeline plugin's retry command
to work.

Instead of using the retry-command provided by the pipeline plugin,
the Jenkinsfile implements this behaviour manually.

Bug: Issue 11927
Change-Id: I419a20f69cabc4c54d7f57fd04ad354563eafd71

Former-commit-id: bf79cd73eb88db2c8cf752e975343051c97bfd54