Show Revert Submission action only on merged changes that have a submission ID
The Revert Submission REST endpoint assumes that every merged change has
a submission ID. A submission ID is only available if the change has
been submitted, however changes can also be set to merged by directly
pushing the patch set commit into the repository (also called
auto-close). If the Revert Submission REST endpoint is invoked for such
a change it fails with:

  java.lang.NullPointerException: merged change 39636 has no submission ID
        at java.util.Objects.requireNonNull(Objects.java:228)
        at com.google.gerrit.server.restapi.change.RevertSubmission.applyImpl(RevertSubmission.java:98)
        at com.google.gerrit.server.restapi.change.RevertSubmission.applyImpl(RevertSubmission.java:50)
        ...

This change is not fixing this issue (Gal is looking into this) but by
not offering the Revert Submission action on changes where it's anyway
failing, we reduce the chances that users hit this issue, hence reducing
the amount of errors in our logs.

Change-Id: I8e76652cce24829a866f5a29836bc699260ba619
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 76485b416e186cfd7bded8207104d92431bad85b