Reduce log level for tests to INFO
The amount of logs that is generated by running the tests is too large
to view it on the CI. Loading a build log in the browser takes so long
that it becomes unusable. This makes it difficult to find the reason for
a failed build/test on the CI.

As this is affecting daily development we need to resolve this quickly,
hence let's reduce the log level for tests to INFO for now.

Having debug logs available to investigate a test failure is good, but
debug logs for passing tests are not useful. Ideally the CI would just
hide debug logs of passing tests, but while the CI doesn't have this
capability we can only reduce the log level for now.

This also affects the log level for running tests locally.

Bug: Issue 11953
Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I6b13ee50459bec18ef1c701fe30c992634ed26d5

Former-commit-id: 134b7e3ece70fe4993619dc034bc126d3aaa21cc