RefControl.ForRefImpl#change(ChangeData): Use ChangeNotes to create ChangeContol
The ChangeNotes are already available from the provided ChangeData, no
need to load them again via ChangeControl.Factory, which happens if we
pass in Change instead of ChangeNotes to create the ChangeControl.

This change made the QueryIT#allChangeOptionsAreServedWithoutExceptions
test fail. This was because:

- this test uses the DETAILED_LABELS option for querying changes
- if the DETAILED_LABELS option is set, ChangeJson may need to test if
  the calling user can remove the reviewers
- this permission check is done on a change, which requires ChangeNotes
  to be loaded
- loading the ChangeNotes failed because lazy loading for the
  DETAILED_LABELS option was not allowed

Before this change the test was passing because the loading of the
ChangeNotes happened outside of ChangeData, but in
ChangeControl.Factory#create which isn't protected by the lazy flag.
This means using the DETAILED_LABELS option isn't possible without
loading the ChangeNotes for the permission check. Marking this option as
requiring lazy load matches reality and makes the test pass. This is
better than lying to ourselves that this option doesn't cause any loads.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Iac176b8e55e9f0e23934c95e0d8402eb55098d1f
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 6942c422b566ad9b0b4807523ed7b269ec1fa887