ReceiveCommitsAdvertiseRefsHook: Advertise the user's open changes
ReceiveCommitsAdvertiseRefsHook works around the fact that git is lacking
negotiation upon push. It adds the last 32 open changes for the repo
to .haves.

This is a heuristical approach. The user is much more likely to have
changes in their local repo that they own rather than the last 32
changes open changes. This commit changes the logic to account for
that.

Change-Id: Ie4261fab9365c2df78d73d5e8d0952f64aa7be38

Former-commit-id: 3a34dc69804cd215c579667b4e46fb8de0b95dc2